requirements.txt
peewee>=3.9.4
platformdirs==3.5.0
python_slugify==8.0.1
